Change package to match maven groupId related to PR4 Fix potential NPE bug in CommandRegistrar commandRolesNotMet if MessageEvent belong to private or group channel. Rename parser.Flag class to parser.CommandFlag due to unpleasant naming conflict with reflection.Flag annotation Refactor multiple minor fragments, mostly code formatting and typos
