+++ src/main/java/com/github/boscojared/disparse/parser/Types.java
@@ -1,0 +1,5 @@
+package com.github.boscojared.disparse.parser;
+
+public enum Types {
+    INT, STR, BOOL, LIST
+}
