+++ src/main/java/com/github/boscojared/disparse/parser/exceptions/NoCommandNameFound.java
@@ -1,0 +1,7 @@
+package com.github.boscojared.disparse.parser.exceptions;
+
+public class NoCommandNameFound extends RuntimeException {
+    public NoCommandNameFound(String message) {
+        super(message);
+    }
+}
