+++ src/main/java/com/github/boscojared/disparse/parser/exceptions/OptionRequired.java
@@ -1,0 +1,5 @@
+package com.github.boscojared.disparse.parser.exceptions;
+
+public class OptionRequired extends RuntimeException {
+    public OptionRequired(String message) { super(message); }
+}
