+++ src/main/java/com/github/boscojared/disparse/parser/exceptions/OptionRequiresValue.java
@@ -1,0 +1,7 @@
+package com.github.boscojared.disparse.parser.exceptions;
+
+public class OptionRequiresValue extends RuntimeException {
+    public OptionRequiresValue(String message) {
+        super(message);
+    }
+}
