+++ src/main/java/com/github/boscojared/disparse/parser/reflection/Flag.java
@@ -1,0 +1,18 @@
+package com.github.boscojared.disparse.parser.reflection;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface Flag {
+    char shortName() default ' ';
+
+    String longName() default "";
+
+    boolean required() default false;
+
+    String description() default "no description available";
+}
