+++ src/main/java/com/github/boscojared/disparse/parser/reflection/Injectable.java
@@ -1,0 +1,11 @@
+package com.github.boscojared.disparse.parser.reflection;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Injectable {
+}
