--- src/main/java/disparse/parser/Types.java
@@ -1,5 +1,0 @@
-package disparse.parser;
-
-public enum Types {
-    INT, STR, BOOL, LIST
-}
