--- src/main/java/disparse/parser/reflection/CommandHandler.java
@@ -1,15 +1,0 @@
-package disparse.parser.reflection;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface CommandHandler {
-    public String commandName();
-    public String description() default "no description available";
-    
-    public String[] roles() default {};
-}
